Added Search By Title Functionality
Added Search By Title Functionality: I added code that allows the user to fetch the Book Title and Author by searching using the title. The user is redirected to another page when viewing the book searched. Lastly, I added a redirect button on the page where the user views the book looked up. The redirect button brings the user back to the main page.