Search By Author Function
Search By Author Function: Created a way for a user to look up a book by the author. Once a user types in an author, the page will redirect the user to a webpage showing the relevant author and book title. Again, there is a button to redirect the user to the mainpage. Please note the input field for looking up by book and author is case sensitive and lookup may be changed to case insensitive in future pushes.