Added Sort By Author and Book Title Functionality
Added Sort By Author and Book Title Functionality: I created two buttons that sort the list of books. One button sorts by author name, and the other button sorts by book title. When a user clicks on either button, the user is taken to another page showing the data sorted based on the button chosen. Again, there is a button that redirects the user back to the main page.